feat: multiusers and analysis and monitoring tools

+++ src/actions/deleteTask.ts
@@ -3,10 +3,16 @@ import { API } from "@/constants";
 import { apiError } from "@/api/apiError";
 import { apiMessages } from "@/api/apiMessages";
 import { ApiResponse, ErrorApi } from "@/api/types";
+import { getUserHashCookie } from "./userHashCookie";
 
 export const deleteTask = async (formData: FormData): Promise<ApiResponse> => {
   try {
+    const userHash = await getUserHashCookie();
     const id = formData.get("id") as string;
+
+    if (!userHash) {
+      throw new Error(apiMessages.error.deleteTask);
+    }
 
     const response = await fetch(`${API}/${id}`, { method: "DELETE" });
     if (!response.ok) throw new Error(apiMessages.error.deleteTask);
